Refactor annotations for python3.6 compatibility
Compatibility with python 3.6 is important since it's the default (and
most up-to-date) version present in RHEL/CentOS 7 and 8. The python 3.6
compatibility requirement is now specified in `mypy.ini`.

Also *flake8* is now configured to only accept type annotations for
non-built-in types in the string literal form, since python 3.6 doesn't
support the `__future__.annotations` module for postponed evaluation of
type annotations.

All the source files have been changed to reflect this new requirement.