Fix preset implementation bug on multi-run-command-benchmarks
Benchmarks which contained multiple run_commands would have their
preset arguments applied to every command instead of the latter only.

Preset arguments are now used to extend the last run_command's arguments
only, as it is considered to be the actual benchmark (stdout from this
command only is used for stat matching).